2024 w40 grass model
A fine bit of grass: This is quite detailed yet makes use of only 32x textures, for 8 new alternatives to the dual grass/stone blocks (or 'Brushrock's, and they're called 'Blushrocks, because they're 'Lush' Brush / Rock nodes) of minetest's RSWO game. Anyway these are the development notes which were the things that kept the project going comfortably.